Implement the Python Source operator using the Rust interface.
