+++ src/python/py_operators/mod.rs
@@ -2,8 +2,10 @@
 mod py_one_in_one_out;
 mod py_one_in_two_out;
 mod py_sink;
+mod py_source;
 
 // Crate-level exports
 pub(crate) use py_one_in_one_out::*;
 pub(crate) use py_one_in_two_out::*;
 pub(crate) use py_sink::*;
+pub(crate) use py_source::*;
